backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
transformers>=4.30.0
torch>=2.0.0
pypdf>=3.0.0
python-docx>=0.8.11
tiktoken>=0.5.0
numpy>=1.24.0
tenacity>=8.0.0
aiosqlite>=0.17.0
redis>=4.0.0

# LangChain and related
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# AI/ML
openai>=1.0.0

# External APIs
wolframalpha>=5.0.0
google-api-python-client>=2.0.0
duckduckgo-search>=4.1.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

# Google Auth
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# New features
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4